extract STT wav conversion into reusable helper
