+++ server/utils/SpeechToText/lemonade/index.js
@@ -1,9 +1,6 @@
-const fs = require("fs");
 const path = require("path");
-const { v4 } = require("uuid");
 const { parseLemonadeServerEndpoint } = require("../../AiProviders/lemonade");
-const { CollectorApi } = require("../../collectorApi");
-const { hotdirPath } = require("../../files");
+const { convertAudioBufferToWav } = require("../helpers");
 
 class LemonadeSTT {
   constructor() {
@@ -28,23 +25,6 @@ class LemonadeSTT {
     console.log(`\x1b[32m[LemonadeSTT]\x1b[0m ${text}`, ...args);
   }
 
-  async #convertToWav(audioBuffer, extension) {
-    const sourceFilename = `stt-${v4()}${extension}`;
-    const sourcePath = path.resolve(hotdirPath, sourceFilename);
-    fs.writeFileSync(sourcePath, audioBuffer);
-
-    const result = await new CollectorApi().convertAudioToWav(sourceFilename);
-    if (!result?.success || !result?.wavFilename)
-      throw new Error(result?.reason || "Audio conversion failed.");
-
-    const wavPath = path.resolve(hotdirPath, result.wavFilename);
-    try {
-      return fs.readFileSync(wavPath);
-    } finally {
-      if (fs.existsSync(wavPath)) fs.unlinkSync(wavPath);
-    }
-  }
-
   /**
    * Transcribes an audio buffer to text using the Lemonade STT service.
    * Whisper backends behind Lemonade reject webm/opus, so any non-wav input
@@ -59,7 +39,7 @@ class LemonadeSTT {
     let payloadBuffer = audioBuffer;
     let payloadFilename = filename;
     if (extension !== ".wav") {
-      payloadBuffer = await this.#convertToWav(audioBuffer, extension);
+      payloadBuffer = await convertAudioBufferToWav(audioBuffer, extension);
       payloadFilename = "audio.wav";
     }
     const file = await toFile(payloadBuffer, payloadFilename);
